+++ packages/fiber/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @react-three/fiber
 
+## 8.6.2
+
+### Patch Changes
+
+- 57c12e9c: fix(types): @react-three/drei declaration files
+
+## 8.6.1
+
+### Patch Changes
+
+- 7a0b5670: fix(core): don't append to unmounted containers
+
 ## 8.6.0
 
 ### Minor Changes